fix (race_archive): split race archive file into 6 files
