+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,9 +1,47 @@
-## Contexte
-<!-- Pourquoi ce PR ? Lien vers le ticket si dispo. -->
+## Summary
+<!-- Short description of the goal of this PR. -->
 
-## Changements
+## Context
+<!-- Why this PR exists. Link the issue/task if available. -->
+
+## Scope
+<!-- What is intentionally included and excluded. -->
+- Included:
+- Not included:
+
+## Changes
+<!-- Concrete technical changes. Keep this factual. -->
 - 
 
-## Vérifications
-- [ ] Tests ajoutés ou mis à jour si nécessaire
-- [ ] Docs mises à jour si nécessaire
+## Impact
+<!-- State what this PR affects. Remove items that do not apply. -->
+- API:
+- Database / Prisma:
+- Provider ingestion:
+- Docs:
+- Scripts / local setup:
+
+## Breaking Changes
+<!-- Write "None" if there are no breaking changes. -->
+
+## Validation
+<!-- List the commands, checks, or manual flows used to validate the PR. -->
+```bash
+# Example:
+# go test ./...
+```
+
+## Review Notes
+<!-- Point reviewers to the highest-risk files, design choices, or tradeoffs. -->
+
+## Rollback Notes
+<!-- If this PR breaks prod/dev, how do we back it out safely? -->
+
+## Checklist
+- [ ] The scope of the PR is clear and limited
+- [ ] Tests were added or updated when needed
+- [ ] Documentation was added or updated when needed
+- [ ] No unrelated changes were included
+- [ ] API changes are documented in `docs/endpoint.md` if applicable
+- [ ] DB / Prisma changes are documented and safe to apply if applicable
+- [ ] I validated the main affected flow locally
